VMPRCI-3071 Use a forked JVM for starting play
As it stands now, sbt is compiled against Scala 2.10. If we attempt to
start a play application that uses a newer version of Scala from
within an sbt plugin, then it will fallover with NoSuchMethodError.

By using a forked JVM, we can run any version of Scala, independent of
the sbt Scala version.